backend/requirements.txt
asgiref==3.5.2
autopep8==2.0.1
backports.zoneinfo==0.2.1
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.4
django-colorfield==0.8.0
django-dotenv==1.4.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
drf-yasg==1.21.4
flake8==6.0.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
isort==5.11.3
itypes==1.2.0
Jinja2==3.1.2
markdown2==2.4.2
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
psycopg2-binary==2.8.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.10.0
PyJWT==2.6.0
pyphen==0.13.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tinycss2==1.2.1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
WeasyPrint==52.5
webcolors==1.12
webencodings==0.5.1
